+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 ## Unreleased - in development
 ### Added
+- gfiAttributes: Adding Boolean type in gfi Attributes so that the original text can be parsed to be more understandable.
 
 ### Changed
 - Coding-Conventions: For unittests in Vue (/src/...) the vast majority of test-folders are called "tests", going back to a mutual understanding of folder naming. Please use "tests" for your unit or e2e tests in Vue in the future.